Clock counter changed to long long

+++ Simulator/TerminalInstrumentation.h
@@ -20,7 +20,7 @@ public:
     // System signals
     sc_in<bool>          i_CLK;
     sc_in<bool>          i_RST;
-    sc_in<unsigned long> i_CLK_CYCLES;
+    sc_in<unsigned long long> i_CLK_CYCLES;
 
     // Link signals
     sc_out<Flit>    o_DATA_OUT;   // Output data channel
